requirements.txt
vnstock
pandas
numpy
pytz
openpyxl
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
anthropic
google-genai
httpx
notebooklm-py
pytest
pytest-asyncio
